requirements.txt
Flask==2.1.3
requests==2.28.1
Pillow==9.2.0
Werkzeug==2.2.1
urllib3==1.26.11
flake8-5.0.1
